Refactor sum calculations and namings
