+++ backend/src/algorithm/utils/class.ts
@@ -1,15 +1,13 @@
-import { range, shuffle } from 'lodash';
+import { range, shuffle, sumBy } from 'lodash';
 import { Class } from '../types';
 import { daysPerWeek } from '../consts';
 
 export function getClassPeriodsPerDay(class_: Class) {
   const workingDays = range(daysPerWeek, 0, -1);
-  let totalPeriodsPerWeek = class_.subjects.reduce((sum, subject) => sum + subject.timesPerWeek, 0);
+  let totalPeriodsPerWeek = sumBy(class_.subjects, 'timesPerWeek');
   const periodsPerDay: number[] = workingDays.map(day => {
-    // calculate average period quantity for that day
-    const periods = Math.trunc(totalPeriodsPerWeek / day);
-    // subtract calculated period quantity from total amount for more equal period distribution
-    totalPeriodsPerWeek -= periods;
+    const periods = Math.trunc(totalPeriodsPerWeek / day); // calculate average period quantity for that day
+    totalPeriodsPerWeek -= periods; // subtract calculated period quantity from total amount for more equal period distribution
     return periods;
   });
